Function oak_abi::channel_read

pub unsafe extern "C" fn channel_read(
    handle: u64, 
    buf: *mut u8, 
    size: usize, 
    actual_size: *mut u32, 
    handle_buf: *mut u8, 
    handle_count: u32, 
    actual_handle_count: *mut u32
) -> u32

Read a message from a channel.

Reads from the channel identified by handle, storing data into buf and associated channel handles into handle_buf. The size of the returned data is indicated by actual_size, and the count of returned channel handles is indicated by actual_handle_count.

If the provided spaces for data (buf plus size) or handles (handle_buf plus 8 x handle_count) are not large enough for the read operation, then no data will be returned and either ErrBufferTooSmall or ErrHandleSpaceTooSmall will be returned. In either case, the required sizes will be returned in the spaces provided by actual_size and actual_handle_count.

Returns the status of the operation, as an OakStatus value. If no message is available on the channel, ErrChannelEmpty will be returned.