Function oak_abi::node_create

pub unsafe extern "C" fn node_create(
    name_buf: *const u8, 
    name_len: usize, 
    config_buf: *const u8, 
    config_len: usize, 
    label_buf: *const u8, 
    label_len: usize, 
    handle: u64
) -> u32

Creates a new Node instance running code identified by a serialized NodeConfiguration.

The name of the new Node is provided in the memory area given by name_buf and name_len. The name does not have to be unique and can be empty. It is used in logs to help identify nodes for debugging purposes.

The serialized configuration object is provided in the memory area given by config_buf and config_len.

The label to assign to the newly created Node is provided in the memory area given by label_buf and label_len.

Returns the status of the operation, as an OakStatus value.