Function oak_abi::wait_on_channels

pub unsafe extern "C" fn wait_on_channels(buf: *mut u8, count: u32) -> u32

Wait for channels to be ready for reading.

Blocks until data is available for reading from one of the specified channel handles. The channel handles are encoded in a buffer buf that holds count contiguous chunks of size SPACE_BYTES_PER_HANDLE.

Returns the status of the operation, as an OakStatus value.