Module oak_abi::label

Labels are defined at the ABI level because they need to be available:

to applications (e.g. via the SDK), so that they can be manipulated by running Oak Nodes

to the Runtime, so that they can be checked and enforced when data is exchanged between Nodes over Channels

In order to support this, Labels are serialized and deserialized across the ABI boundary, and can be read and write using ABI functions.

Modules

tag

Structs

GrpcTag
Policies related to gRPC communication, referring to the native gRPC node within the TCB.

Label
Label represents a policy associated with a message or a Node.

Tag
Tag represents a category of secrecy or integrity that is associated with data within Oak, and refers to a Node or family of Nodes which are able to declassify data with that tag.

TlsEndpointTag
Policies related to HTTPS communication.

WebAssemblyModuleTag
Policies related to modules, referring to the native WebAssembly node within the TCB.

Functions

authorization_bearer_token_hmac_tag
Creates a Tag having as principal the provided authorization bearer token.