Module oak_abi::proto::oak::application

Modules

node_configuration

Structs

ApplicationConfiguration
An ApplicationConfiguration represents a unit of deployment in Oak.

ConfigMap
A serialized list of key-value pairs that are specified as command line flags to the Oak Loader binary, and are made available to the initial Node of the running Oak Application.

GrpcClientConfiguration
GrpcClientConfiguration describes the configuration of a gRPC client pseudo-Node (which is provided by the Oak Runtime), connected to a specific external (non-Oak) gRPC service.

GrpcServerConfiguration
GrpcServerConfiguration describes the configuration of a gRPC server pseudo-Node (which is provided by the Oak Runtime), that processes gRPC requests from external (non-Oak) clients.

LogConfiguration
LogConfiguration describes the configuration of a logging pseudo-Node (which is provided by the Oak Runtime).

NodeConfiguration
NodeConfiguration indicates the configuration of a created Node.

RoughtimeClientConfiguration
RoughtimeClientConfiguration describes the configuration of a Roughtime client pseudo-Node (which is provided by the Oak Runtime), with the given external Roughtime servers and connection parameters.

RoughtimeServer
Information to identify a particular Roughtime server. Only UDP and Ed25519 public keys are currently supported.

StorageProxyConfiguration
StorageProxyConfiguration describes the configuration of a storage proxy pseudo-Node (which is provided by the Oak Runtime), connected to a specific storage provider.

WebAssemblyConfiguration
WebAssemblyConfiguration describes the configuration of a Web Assembly based Node.