Crate oak_abi

Type, constant and Wasm host function definitions for the Oak application binary interface (ABI).

Re-exports

pub use proto::oak::ChannelReadStatus;

pub use proto::oak::OakStatus;

Modules

grpc

label
Labels represent the kinds of information that is allowed to be processed by entities in the Oak system. The Oak Runtime allows Oak Nodes to manipulate labels, and so the labels need to be passed across the Oak ABI in a defined binary format. That format is a serialized protocol buffer holding the Label message defined in the policy.proto file.

proto

Constants

INVALID_HANDLE
Invalid handle value.

SPACE_BYTES_PER_HANDLE
Number of bytes needed per-handle for channel readiness notifications.

Functions

channel_close⚠
Closes the channel identified by handle.

channel_create⚠
Create a new unidirectional Channel.

channel_read⚠
Read a message from a channel.

channel_write⚠
Write a message to a channel.

node_create⚠
Creates a new Node instance running code identified by a serialized NodeConfiguration.

random_get⚠
Fill a buffer with random data.

wait_on_channels⚠
Wait for channels to be ready for reading.

Type Definitions

Handle
Handle used to identify read or write channel halves.

NodeMain
Expected Rust type for a Node entrypoint.

NodeMainC
Expected type for a Node entrypoint that is exposed as a Wasm export.