Module oak_abi::label

Labels represent the kinds of information that is allowed to be processed by entities in the Oak system. The Oak Runtime allows Oak Nodes to manipulate labels, and so the labels need to be passed across the Oak ABI in a defined binary format. That format is a serialized protocol buffer holding the Label message defined in the policy.proto file.

Re-exports

pub use crate::proto::oak::label::*;

Functions

authorization_bearer_token_hmac_tag
Creates a Tag having as principal the provided authorization bearer token.

web_assembly_module_tag
Creates a Tag having as principal the provided WebAssembly module SHA-256 hash.